feat: convert JSX components to TSX
