+++ apps/desktop/app/ScrollSectionsWrapper.tsx
@@ -1,11 +1,15 @@
-// ScrollSectionsWrapper.js
+// ScrollSectionsWrapper.tsx
 "use client";
 
-import React from "react";
+import type { ReactNode } from "react";
 import { useScrollAnchors } from "@src/features/navigation/utils/scrollUtils";
 import { sections } from "@src/features/navigation/data/sections";
 
-const ScrollSectionsWrapper = ({ children }) => {
+type ScrollSectionsWrapperProps = {
+    children: ReactNode;
+};
+
+const ScrollSectionsWrapper = ({ children }: ScrollSectionsWrapperProps) => {
     useScrollAnchors(sections);
     return <>{children}</>;
 };
