+++ apps/desktop/app/headerProps.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { NavigationProvider } from "@utils/context/NavigationContext";
+import { Header } from "@src/features/navigation/index";
+
+const HeaderProps = () => {
+    return (
+        <NavigationProvider>
+            <Header />
+        </NavigationProvider>
+    );
+};
+
+export default HeaderProps;
