+++ apps/desktop/src/features/navigation/components/Header.tsx
@@ -1,29 +1,17 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import Nav from "./Nav";
 import Logo from "@src/features/navigation/icons/Logo";
 import { useScrollContext } from "@utils/context/ScrollContext";
 import { useNavigation } from "@utils/context/NavigationContext";
-import { MenuItem, menuItems } from "@src/features/navigation/data/menuItems";
+import { menuItems } from "@src/features/navigation/data/menuItems";
 import { updateMenuClasses } from "@src/features/navigation/utils/updateMenuUtils";
 import { handleScrollClick, handleNavClick } from "@src/features/navigation/utils/fnScrollUtils";
 import { useInitialScroll } from "@src/features/navigation/utils/scrollUtils";
 import useResize from "@src/features/navigation/components/utils/useResize";
 
-interface NavProps {
-    menuItems: MenuItem[];
-    onNavigationClick: (path: string, scrollOffset?: number) => void;
-    openButton: boolean;
-    openMainButton: boolean;
-    tabletMain: boolean;
-    bigMenu: boolean;
-    setBigMenu: React.Dispatch<React.SetStateAction<boolean>>;
-    setOpenMainButton: React.Dispatch<React.SetStateAction<boolean>>;
-    setTabletMain: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-const Header: React.FC<NavProps> = () => {
+const Header = () => {
     const pathname = usePathname();
     const { currentRoute, updateRoute } = useNavigation();
     const { activeSection } = useScrollContext();
