+++ apps/desktop/src/features/navigation/icons/Blog.tsx
@@ -1,0 +1,14 @@
+import { memo } from "react";
+
+const Blog = () => (
+    <div className="nav-icon">
+    <svg fill="none" aria-hidden="true" viewBox="0 0 18 18">
+      <path
+        d="M2.94609 8.25L6.69609 10.05H3.84609C4.14609 10.65 4.59609 11.25 4.89609 11.7L9.99609 13.5H6.39609C7.59609 14.55 9.24609 15.3 11.0461 15.45C12.5461 15.6 14.0461 15.75 15.5461 15.75L17.4961 17.7C17.7961 18 18.3961 18 18.6961 17.7C18.9961 17.4 18.9961 16.8 18.6961 16.5L9.69609 7.5C9.54609 7.35 9.54609 6.9 9.69609 6.75C9.84609 6.6 10.2961 6.6 10.4461 6.75L16.7461 13.05C16.7461 12.15 16.5961 11.1 16.5961 10.2C15.5461 1.35 2.79609 0.15 0.996094 0C0.996094 0.9 1.44609 4.65 2.94609 8.25Z"
+        className="icon-color"
+      />
+    </svg>
+    </div>
+);
+
+export default memo(Blog);
