--- apps/desktop/src/features/navigation/icons/Connection.jsx
@@ -1,19 +1,0 @@
-import React from "react";
-
-const Connection = () => {
-  return (
-    <div className="nav-icon">
-      <svg fill="none" aria-hidden="true" viewBox="0 0 18 18">
-        <path
-          d="M8.94406 9.71053C11.6314 9.71053 13.81 7.53675 13.81 4.85526C13.81 2.17378 11.6314 0 8.94406 0C6.25668 0 4.07812 2.17378 4.07812 4.85526C4.07812 7.53675 6.25668 9.71053 8.94406 9.71053Z"
-          className="icon-color"
-        />
-        <path
-          d="M13.6105 10.6578H11.4742C10.8808 10.9736 9.73355 11.171 9.02146 11.171C8.26981 11.171 7.16212 10.9736 6.56872 10.6578H4.35333C2.81047 10.6578 1.62366 11.6052 1.10938 12.9078C2.49399 15.9078 5.50058 17.9999 9.02146 17.9999C12.5423 17.9999 15.5885 15.9078 16.9336 12.9078C16.3006 11.5657 15.1533 10.6578 13.6105 10.6578Z"
-          className="icon-color"
-        />
-      </svg>
-    </div>
-  );
-};
-export default React.memo(Connection);
