--- apps/desktop/src/features/navigation/icons/Contact.jsx
@@ -1,21 +1,0 @@
-import React from "react";
-
-const Contact = () => {
-  return (
-    <div className="nav-icon">
-      <svg
-        width="18"
-        height="18"
-        fill="none"
-        aria-hidden="true"
-        viewBox="0 0 18 18"
-      >
-        <path
-          d="M13.0752 6.99287C13.0752 4.27344 10.0968 1.9425 6.60041 1.9425C3.10401 1.9425 -0.00390625 4.14394 -0.00390625 6.99287C-0.00390625 8.02884 0.384583 9.06481 1.16156 9.84179C0.773072 10.7483 0.12559 11.5252 0.12559 11.5252C-0.00390625 11.6547 -0.00390625 11.6547 -0.00390625 11.7842C-0.00390625 11.9137 0.12559 11.9137 0.255087 11.9137C1.42055 11.9137 2.32703 11.5252 2.97451 11.1368C4.01048 11.6547 5.17595 11.9137 6.47091 11.9137C10.0968 11.9137 13.0752 9.71229 13.0752 6.99287ZM16.8306 13.8562C17.6076 13.0792 17.9961 12.0432 17.9961 11.0073C17.9961 8.93531 16.3126 7.12236 13.9817 6.34538C13.9817 6.60438 13.9817 6.73387 13.9817 6.99287C13.9817 10.2303 10.6148 12.9497 6.47091 12.9497C6.08243 12.9497 5.82343 12.9497 5.43494 12.9497C6.47091 14.7626 8.80185 16.0576 11.3918 16.0576C12.6867 16.0576 13.8522 15.7986 14.8882 15.2806C15.5357 15.6691 16.5716 16.0576 17.6076 16.0576C17.7371 16.0576 17.7371 16.0576 17.8666 15.9281C17.8666 15.7986 17.8666 15.6691 17.8666 15.6691C17.9961 15.5396 17.2191 14.7626 16.8306 13.8562Z"
-          className="icon-color"
-        />
-      </svg>
-    </div>
-  );
-};
-export default React.memo(Contact);
