--- apps/desktop/src/features/navigation/icons/Reservation.jsx
@@ -1,186 +1,0 @@
-import React from "react";
-
-const Reservation = () => {
-  return (
-    <div className="nav-icon">
-      <svg fill="none" aria-hidden="true" viewBox="0 0 18 18">
-        <g clipPath="url(#clip0_167_8479)">
-          <g opacity="0.5" filter="url(#filter0_d_167_8479)">
-            <path
-              d="M15.3008 6.30005H2.70078C1.80078 6.30005 1.80078 7.00005 1.80078 7.20005V14.4C1.80078 15.3 2.72078 16.2 3.60078 16.2H14.4008C15.3208 16.2 16.2008 15.3 16.2008 14.4V7.20005C16.2008 7.00005 16.2008 6.30005 15.3008 6.30005Z"
-              className="icon-color_secondary"
-            />
-          </g>
-          <g filter="url(#filter1_d_167_8479)">
-            <path
-              d="M16.22 1.8H15.3V0.45C15.3 0.15 15 0 14.85 0H13.05C12.9 0 12.6 0.15 12.6 0.45V1.8H5.4V0.45C5.4 0.15 5.1 0 4.95 0H3.15C3 0 2.7 0.15 2.7 0.45V1.8H1.79C0.81 1.8 0 2.61 0 3.59V4.51C0 5 0.4 5.4 0.88 5.4H17.11C17.6 5.4 17.99 5 17.99 4.52V3.59C18 2.6 17.2 1.8 16.22 1.8Z"
-              fill="white"
-            />
-          </g>
-          <g filter="url(#filter2_d_167_8479)">
-            <path
-              d="M7.47445 12.9117H6.52068V7.54784C6.52068 7.3535 6.41363 7.16888 6.25791 7.07171C6.09246 6.97454 5.88808 6.97454 5.72263 7.08142L4.25304 8.0337C4.13625 8.11144 4.04866 8.22804 4.01946 8.3738C3.99027 8.51955 4.00973 8.66531 4.07786 8.78192C4.22384 9.03456 4.54501 9.1123 4.79805 8.95682L5.47932 8.51955V12.9117H4.52555C4.23358 12.9117 4 13.1546 4 13.4558C4 13.7571 4.23358 14 4.52555 14H7.47445C7.76642 14 8 13.7571 8 13.4558C8 13.1546 7.76642 12.9117 7.47445 12.9117Z"
-              fill="white"
-            />
-            <path
-              d="M13.47 12.9367H10.89L12.95 10.9578C13.4 10.5246 13.66 9.91421 13.66 9.28411C13.66 8.02391 12.66 7 11.43 7H11.32C10.7 7 10.1 7.25598 9.67 7.69902L9.16 8.22082C9.06 8.31927 9 8.46695 9 8.60478C9 8.75246 9.06 8.8903 9.16 8.98875C9.26 9.09705 9.39 9.14627 9.54 9.14627C9.68 9.14627 9.82 9.0872 9.92 8.98875L10.43 8.4571C10.67 8.21097 10.99 8.07314 11.32 8.07314H11.43C12.07 8.07314 12.59 8.60478 12.59 9.26442C12.59 9.58931 12.46 9.91421 12.22 10.1308L9.17 13.0549C9.01 13.2124 8.96 13.4487 9.04 13.6554C9.12 13.8622 9.32 14 9.54 14H13.47C13.76 14 14 13.7539 14 13.4487C14 13.1435 13.76 12.9367 13.47 12.9367Z"
-              fill="white"
-            />
-          </g>
-          <g filter="url(#filter3_d_167_8479)">
-            <path
-              d="M0 3.59998V17.1C0 17.6 0.4 18 0.9 18H17.06C17.58 18 17.99 17.58 17.99 17.07V3.59998H0ZM16.6 17.1H1.31C1.08 17.1 0.9 16.91 0.9 16.69V4.49998H17.1V16.6C17.1 16.88 16.88 17.1 16.6 17.1Z"
-              fill="white"
-            />
-          </g>
-        </g>
-        <defs>
-          <filter
-            id="filter0_d_167_8479"
-            x="-2.19922"
-            y="6.30005"
-            width="22.3984"
-            height="17.9"
-            filterUnits="userSpaceOnUse"
-            colorInterpolationFilters="sRGB"
-          >
-            <feFlood floodOpacity="0" result="BackgroundImageFix" />
-            <feColorMatrix
-              in="SourceAlpha"
-              type="matrix"
-              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              result="hardAlpha"
-            />
-            <feOffset dy="4" />
-            <feGaussianBlur stdDeviation="2" />
-            <feComposite in2="hardAlpha" operator="out" />
-            <feColorMatrix
-              type="matrix"
-              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-            />
-            <feBlend
-              mode="normal"
-              in2="BackgroundImageFix"
-              result="effect1_dropShadow_167_8479"
-            />
-            <feBlend
-              mode="normal"
-              in="SourceGraphic"
-              in2="effect1_dropShadow_167_8479"
-              result="shape"
-            />
-          </filter>
-          <filter
-            id="filter1_d_167_8479"
-            x="-4"
-            y="0"
-            width="25.9883"
-            height="13.4"
-            filterUnits="userSpaceOnUse"
-            colorInterpolationFilters="sRGB"
-          >
-            <feFlood floodOpacity="0" result="BackgroundImageFix" />
-            <feColorMatrix
-              in="SourceAlpha"
-              type="matrix"
-              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              result="hardAlpha"
-            />
-            <feOffset dy="4" />
-            <feGaussianBlur stdDeviation="2" />
-            <feComposite in2="hardAlpha" operator="out" />
-            <feColorMatrix
-              type="matrix"
-              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-            />
-            <feBlend
-              mode="normal"
-              in2="BackgroundImageFix"
-              result="effect1_dropShadow_167_8479"
-            />
-            <feBlend
-              mode="normal"
-              in="SourceGraphic"
-              in2="effect1_dropShadow_167_8479"
-              result="shape"
-            />
-          </filter>
-          <filter
-            id="filter2_d_167_8479"
-            x="0"
-            y="7"
-            width="18"
-            height="15"
-            filterUnits="userSpaceOnUse"
-            colorInterpolationFilters="sRGB"
-          >
-            <feFlood floodOpacity="0" result="BackgroundImageFix" />
-            <feColorMatrix
-              in="SourceAlpha"
-              type="matrix"
-              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              result="hardAlpha"
-            />
-            <feOffset dy="4" />
-            <feGaussianBlur stdDeviation="2" />
-            <feComposite in2="hardAlpha" operator="out" />
-            <feColorMatrix
-              type="matrix"
-              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-            />
-            <feBlend
-              mode="normal"
-              in2="BackgroundImageFix"
-              result="effect1_dropShadow_167_8479"
-            />
-            <feBlend
-              mode="normal"
-              in="SourceGraphic"
-              in2="effect1_dropShadow_167_8479"
-              result="shape"
-            />
-          </filter>
-          <filter
-            id="filter3_d_167_8479"
-            x="-4"
-            y="3.59998"
-            width="25.9883"
-            height="22.4"
-            filterUnits="userSpaceOnUse"
-            colorInterpolationFilters="sRGB"
-          >
-            <feFlood floodOpacity="0" result="BackgroundImageFix" />
-            <feColorMatrix
-              in="SourceAlpha"
-              type="matrix"
-              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              result="hardAlpha"
-            />
-            <feOffset dy="4" />
-            <feGaussianBlur stdDeviation="2" />
-            <feComposite in2="hardAlpha" operator="out" />
-            <feColorMatrix
-              type="matrix"
-              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-            />
-            <feBlend
-              mode="normal"
-              in2="BackgroundImageFix"
-              result="effect1_dropShadow_167_8479"
-            />
-            <feBlend
-              mode="normal"
-              in="SourceGraphic"
-              in2="effect1_dropShadow_167_8479"
-              result="shape"
-            />
-          </filter>
-          <clipPath id="clip0_167_8479">
-            <rect width="18" height="18" fill="white" />
-          </clipPath>
-        </defs>
-      </svg>
-    </div>
-  );
-};
-export default React.memo(Reservation);
