+++ apps/desktop/src/features/navigation/icons/Search.tsx
@@ -1,0 +1,20 @@
+import { memo } from "react";
+
+const Search = () => (
+    <span className="nav-icon">
+    <svg
+      width="18"
+      height="18"
+      fill="none"
+      aria-hidden="true"
+      viewBox="0 0 18 18"
+    >
+      <path
+        d="M17.7 15.6L14.25 12C14.1 11.85 13.95 11.85 13.65 11.85H13.05C13.95 10.65 14.55 9 14.55 7.35C14.7 3.3 11.4 0 7.35 0C3.3 0 0 3.3 0 7.35C0 11.4 3.3 14.7 7.35 14.7C9 14.7 10.65 14.1 11.85 13.2V13.8C11.85 14.1 12 14.25 12.15 14.4L15.6 17.85C15.9 18.15 16.5 18.15 16.8 17.85L17.85 16.8C18.15 16.5 18.15 15.9 17.7 15.6ZM7.35 11.85C4.8 11.85 2.85 9.9 2.85 7.35C2.85 4.8 4.8 2.85 7.35 2.85C9.9 2.85 11.85 4.8 11.85 7.35C11.85 9.75 9.75 11.85 7.35 11.85Z"
+        className="icon-color"
+      />
+    </svg>
+    </span>
+);
+
+export default memo(Search);
