--- apps/desktop/src/features/navigation/icons/Services.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-
-const Services = () => {
-  return (
-    <div className="nav-icon">
-      <svg fill="none" aria-hidden="true" viewBox="0 0 18 18">
-        <path
-          d="M2.84609 12.9H0.596094C0.296094 12.9 -0.00390625 13.2 -0.00390625 13.5V15.75C-0.00390625 16.05 0.296094 16.35 0.596094 16.35H2.84609C3.14609 16.35 3.44609 16.05 3.44609 15.75V13.5C3.29609 13.2 3.14609 12.9 2.84609 12.9ZM2.84609 1.65002H0.596094C0.296094 1.65002 -0.00390625 1.95002 -0.00390625 2.25002V4.50002C-0.00390625 4.80002 0.296094 5.10002 0.596094 5.10002H2.84609C3.14609 5.10002 3.44609 4.80002 3.44609 4.50002V2.25002C3.29609 1.95002 3.14609 1.65002 2.84609 1.65002ZM2.84609 7.35002H0.596094C0.296094 7.35002 -0.00390625 7.50002 -0.00390625 7.80002V10.05C-0.00390625 10.35 0.296094 10.65 0.596094 10.65H2.84609C3.14609 10.65 3.44609 10.35 3.44609 10.05V7.80002C3.29609 7.50002 3.14609 7.35002 2.84609 7.35002ZM17.3961 13.5H6.14609C5.84609 13.5 5.69609 13.65 5.69609 13.95V15.15C5.69609 15.45 5.99609 15.75 6.29609 15.75H17.5461C17.6961 15.75 17.9961 15.45 17.9961 15.15V13.95C17.9961 13.65 17.6961 13.5 17.3961 13.5ZM17.3961 2.25002H6.14609C5.84609 2.25002 5.54609 2.55002 5.54609 2.85002V4.05002C5.54609 4.35002 5.84609 4.65002 6.14609 4.65002H17.3961C17.6961 4.65002 17.9961 4.35002 17.9961 4.05002V2.85002C17.9961 2.40002 17.6961 2.25002 17.3961 2.25002ZM17.3961 7.80002H6.14609C5.84609 7.80002 5.54609 8.10002 5.54609 8.40002V9.60002C5.54609 9.90002 5.84609 10.2 6.14609 10.2H17.3961C17.6961 10.2 17.9961 9.90002 17.9961 9.60002V8.40002C17.9961 8.10002 17.6961 7.80002 17.3961 7.80002Z"
-          className="icon-color"
-        />
-      </svg>
-    </div>
-  );
-};
-export default React.memo(Services);
