+++ apps/desktop/src/features/navigation/icons/Tarifs.tsx
@@ -1,0 +1,14 @@
+import { memo } from "react";
+
+const Tarifs = () => (
+    <div className="nav-icon">
+    <svg viewBox="0 0 18 18" fill="none" aria-hidden="true">
+      <path
+        d="M16.313 9A7.31 7.31 0 0 0 9 1.688c-1.939 0-3.799.77-5.171 2.142S1.688 7.061 1.688 9A7.31 7.31 0 0 0 9 16.313 7.31 7.31 0 0 0 16.313 9zM0 9a9 9 0 0 1 9-9 9 9 0 0 1 9 9 9 9 0 0 1-9 9 9 9 0 0 1-9-9zm4.5.563h.563c-.021-.176-.032-.355-.032-.534a4.67 4.67 0 0 1 .039-.591H4.5c-.309 0-.562-.253-.562-.562s.253-.562.563-.562h.868A4.53 4.53 0 0 1 9.559 4.5h1.445a.84.84 0 1 1 0 1.688H9.559c-.925 0-1.744.443-2.264 1.125h2.83c.309 0 .563.253.563.563s-.253.563-.562.563H6.782a2.95 2.95 0 0 0-.06.591 2.96 2.96 0 0 0 .049.534h3.354c.309 0 .563.253.563.563s-.253.563-.562.563H7.256c.517.714 1.354 1.178 2.303 1.178h1.445a.84.84 0 1 1 0 1.688H9.559a4.53 4.53 0 0 1-4.212-2.865H4.5c-.309 0-.562-.253-.562-.562s.253-.562.563-.562z"
+        className="icon-color"
+      />
+    </svg>
+    </div>
+);
+
+export default memo(Tarifs);
