+++ apps/desktop/src/features/navigation/icons/social/Instagram.tsx
@@ -1,0 +1,15 @@
+import { memo } from "react";
+const Instagram = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 90 90">
+    <path
+      d="M58.7 18.6H31.3c-7 0-12.7 5.7-12.7 12.7v27.5c0 7 5.7 12.7 12.7 12.7h27.5c7 0 12.7-5.7 12.7-12.7V31.3c-.1-7-5.8-12.7-12.8-12.7zM45 61.9c-9.3 0-16.9-7.6-16.9-16.9S35.7 28.1 45 28.1 61.9 35.7 61.9 45 54.3 61.9 45 61.9zm16.6-29.6a3.8 3.8 0 1 1 3.8-3.8c-.1 2.1-1.7 3.8-3.8 3.8zM45 34.5a10.5 10.5 0 1 0 0 21 10.5 10.5 0 1 0 0-21zM0 0v90h90V0H0zm77.9 58.7c0 10.6-8.6 19.1-19.1 19.1H31.3c-10.6 0-19.1-8.6-19.1-19.1V31.3c0-10.6 8.6-19.1 19.1-19.1h27.5c10.6 0 19.1 8.6 19.1 19.1v27.4z"
+      className="bg-color"
+    />
+    <g className="icon-color">
+      <path d="M58.7 77.9H31.3c-10.6 0-19.1-8.6-19.1-19.1V31.3c0-10.6 8.6-19.1 19.1-19.1h27.5c10.6 0 19.1 8.6 19.1 19.1v27.5c0 10.5-8.6 19.1-19.2 19.1zM31.3 18.6c-7 0-12.7 5.7-12.7 12.7v27.5c0 7 5.7 12.7 12.7 12.7h27.5c7 0 12.7-5.7 12.7-12.7V31.3c0-7-5.7-12.7-12.7-12.7H31.3zM45 61.9c-9.3 0-16.9-7.6-16.9-16.9S35.7 28.1 45 28.1 61.9 35.7 61.9 45 54.3 61.9 45 61.9zm0-27.4a10.5 10.5 0 1 0 0 21 10.5 10.5 0 1 0 0-21z" />
+      <circle cx="61.6" cy="28.5" r="3.8" />
+    </g>
+    </svg>
+);
+
+export default memo(Instagram);
