+++ apps/desktop/src/features/navigation/icons/social/Linkedin.tsx
@@ -1,0 +1,15 @@
+import { memo } from "react";
+const Linkedin = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 90 90">
+    <path
+      d="M90 6.6v76.9c0 3.6-2.9 6.6-6.6 6.6H6.6C3 90 0 87.1 0 83.4V6.6C0 2.9 3 0 6.6 0h76.9C87.1 0 90 2.9 90 6.6h0zM26.3 33.8H13.2v43.1h13.1V33.8zm1.1-14.1c.1-4.2-3.3-7.7-7.6-7.8-4.2-.1-7.7 3.3-7.8 7.6-.1 4.2 3.3 7.7 7.6 7.8h.1c4.1.1 7.6-3.2 7.7-7.3v-.3zM76.9 51c0-13.1-8.7-18.2-16.2-18.2s-13.1 4.4-14.4 7h-.2v-6.1H33.8v43.1h13.1V54.4c0-6.9 4.7-10 9.3-10 3.8 0 7.6 2.9 7.6 9.8v22.7h13.1V51z"
+      className="bg-color"
+    />
+    <path
+      d="M76.9 76.9V51c0-13.1-8.7-18.2-16.2-18.2s-13.1 4.4-14.4 7h-.2v-6.1H33.8v43.1h13.1V54.4c0-6.9 4.7-10 9.3-10 3.8 0 7.6 2.9 7.6 9.8v22.7h13.1zM27.4 20v-.3c.1-4.2-3.3-7.7-7.6-7.8-4.2-.1-7.7 3.3-7.8 7.6-.1 4.2 3.3 7.7 7.6 7.8h.1c4.2.1 7.6-3.2 7.7-7.3zM13.2 33.8h13.1v43.1H13.2z"
+      className="icon-color"
+    />
+    </svg>
+);
+
+export default memo(Linkedin);
