--- apps/desktop/src/features/navigation/icons/social/TikTok.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-const TikTok = () => {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 90 90">
-      <path
-        d="M40.5 68.9zM0 0v90h90V0H0zm73 39.3c-5.5 0-10.9-1.7-15.4-4.9v22.5c0 11.2-9.1 20.3-20.3 20.3-4.2 0-8.2-1.3-11.6-3.7-5.4-3.8-8.7-10-8.7-16.7 0-11.2 9.1-20.3 20.3-20.3.9 0 1.8.1 2.7.2v2.6h0 0V48c-.7-.2-1.5-.4-2.2-.4h0-.5c-5.1 0-9.3 4.2-9.3 9.3 0 3.5 1.9 6.6 5 8.2h0a9.16 9.16 0 0 0 4.3 1.1c5 0 9.1-4 9.3-8.9V13h11a13.33 13.33 0 0 0 .3 2.8c.8 4.1 3.2 7.8 6.7 10.1 2.5 1.6 5.4 2.5 8.4 2.5v2.5 8.4z"
-        className="bg-color"
-      />
-      <path
-        d="M73 30.7v-2.5c-3 0-5.9-.9-8.4-2.5-3.5-2.3-6-5.9-6.7-10.1a13.33 13.33 0 0 1-.3-2.8h-11v44.3c-.2 5-4.3 8.9-9.3 8.9a9.16 9.16 0 0 1-4.3-1.1h0a9.25 9.25 0 0 1-5-8.2c0-5.1 4.2-9.3 9.3-9.3h.5 0c.8 0 1.5.2 2.2.4v-8.7h0 0v-2.6c-.9-.1-1.8-.2-2.7-.2-11.2.2-20.3 9.3-20.3 20.5 0 6.6 3.2 12.9 8.7 16.7 3.4 2.4 7.5 3.7 11.6 3.7 11.2 0 20.3-9.1 20.3-20.3V34.3c4.5 3.2 9.9 4.9 15.4 4.9v-8.5zM40.5 68.9z"
-        className="icon-color"
-      />
-    </svg>
-  );
-};
-export default React.memo(TikTok);
