+++ apps/desktop/src/features/navigation/icons/social/WhatsApp.tsx
@@ -1,0 +1,15 @@
+import { memo } from "react";
+const WhatsApp = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 90 90">
+    <path
+      d="M0 0v90h90V0H0zm74.7 55.8c-1.5 3.9-3.7 7.5-6.6 10.6A31.16 31.16 0 0 1 58 73.6c-3.8 1.7-7.9 2.7-12.1 2.8H45c-4.8 0-9.5-1.1-13.8-3.1L14.6 77h0-.1s-.1 0-.1-.1v-.1-.1L17.1 60c-2.6-4.8-4-10.3-3.9-15.8s1.6-10.9 4.4-15.6 6.8-8.7 11.6-11.4C34.1 14.4 39.5 13 45 13h.2c8.2 0 16 3.3 21.9 9s9.3 13.4 9.6 21.6c.2 4.1-.5 8.3-2 12.2zM45 18.5h-.8c-4.6.1-9 1.4-12.9 3.8S24.2 28 22 32s-3.3 8.5-3.3 13c.1 4.6 1.3 9 3.6 13l.5.8-2.1 11.4L32 67.5l.8.4c3.7 1.9 7.9 2.9 12.1 2.9h.8 0c6.9-.2 13.4-3.1 18.1-8.1 4.8-5 7.4-11.6 7.3-18.5s-2.9-13.4-7.8-18.3c-4.8-4.7-11.4-7.4-18.3-7.4zm15.1 37.8c-.7 1.8-4 3.5-5.5 3.6-.4 0-.8.1-1.4.1-1.3 0-3.4-.3-8.1-2.4-8-3.4-12.9-11.9-13.2-12.4-.4-.5-3.1-4.4-3-8.3s2.2-5.8 3-6.6c.3-.3.6-.5.9-.7.4-.2.7-.2 1.1-.3h0 .1c.5 0 1.1 0 1.5.1.6 0 1.2.1 1.7 1.5.7 1.6 2.1 5.6 2.3 6 .1.2.2.5.2.7s-.1.5-.2.7c-.2.5-.5.9-.8 1.3-.4.5-.8 1-1.2 1.4s-.8.8-.4 1.6c1.1 2.1 2.6 4 4.3 5.7 1.8 1.8 3.9 3.2 6.3 4.2.3.1.6.2.9.2.2 0 .3 0 .5-.1s.3-.2.4-.3c.5-.5 2-2.2 2.6-3a.78.78 0 0 1 .4-.4c.2-.1.4-.2.5-.2.3 0 .6.1.9.2.7.3 4.6 2.4 5.4 2.8s1.3.6 1.5 1c.1.2 0 1.8-.7 3.6z"
+      className="bg-color"
+    />
+    <path
+      d="M76.7 43.6c-.3-8.2-3.7-15.9-9.6-21.6a32 32 0 0 0-21.9-9H45c-5.5 0-10.9 1.4-15.7 4.2-4.8 2.7-8.8 6.7-11.6 11.4s-4.3 10.1-4.4 15.6 1.2 11 3.9 15.8l-2.8 16.8v.1.1s.1 0 .1.1.1 0 .1 0h0l16.6-3.7c4.3 2.1 9 3.1 13.8 3.1h.9c4.2-.1 8.3-1.1 12.1-2.8s7.3-4.2 10.1-7.2c2.9-3.1 5.1-6.7 6.6-10.6 1.5-4 2.2-8.2 2-12.3zM45.8 70.9H45c-4.2 0-8.4-1-12.1-2.9l-.8-.4-11.3 2.7 2.1-11.4-.5-.9c-2.3-3.9-3.5-8.4-3.6-13s1.1-9.1 3.3-13c2.2-4 5.4-7.3 9.3-9.7s8.3-3.7 12.9-3.8h.8c6.9 0 13.5 2.7 18.4 7.6 4.9 4.8 7.7 11.4 7.8 18.3S68.8 57.9 64 62.9c-4.8 4.9-11.3 7.8-18.2 8h0zM34 29.4c-.4 0-.8.1-1.1.3s-.7.4-.9.7c-.8.8-2.9 2.6-3 6.6-.1 3.9 2.6 7.8 3 8.3s5.2 9 13.2 12.4c4.7 2 6.8 2.4 8.1 2.4.6 0 1-.1 1.4-.1 1.5-.1 4.8-1.8 5.5-3.6s.8-3.5.6-3.8-.7-.6-1.5-1-4.6-2.5-5.4-2.8c-.3-.1-.6-.2-.9-.2-.2 0-.4.1-.5.2a.78.78 0 0 0-.4.4c-.6.8-2.1 2.5-2.6 3-.1.1-.2.2-.4.3a.9.9 0 0 1-.5.1c-.3 0-.6-.1-.9-.2-2.3-1-4.5-2.4-6.3-4.2-1.7-1.7-3.1-3.6-4.3-5.7-.4-.8 0-1.2.4-1.6s.8-.9 1.2-1.4c.3-.4.6-.8.8-1.3.1-.2.2-.5.2-.7s-.1-.5-.2-.7c-.2-.4-1.6-4.4-2.3-6-.5-1.4-1.2-1.4-1.7-1.5-.4.1-.9.1-1.5.1h0"
+      className="icon-color"
+    />
+    </svg>
+);
+
+export default memo(WhatsApp);
