--- apps/desktop/src/features/navigation/icons/social/Youtube.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-const Youtube = () => {
-  return (
-    <svg viewBox="0 0 90 90" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M55.2 45.1l-17-9.6v19.3l17-9.7zM0 .1v90h90V.1H0zm76.5 60.6c-.8 2.7-2.9 5-5.8 5.7-5.1 1.4-25.7 1.4-25.7 1.4s-20.6 0-25.7-1.4c-2.8-.8-5-3-5.8-5.7-1.4-5-1.4-15.6-1.4-15.6s0-10.5 1.4-15.5c.8-2.8 3-5 5.8-5.7 5.1-1.4 25.7-1.4 25.7-1.4s20.6 0 25.7 1.4c2.9.7 5 2.9 5.8 5.7 1.4 5 1.4 15.5 1.4 15.5s0 10.6-1.4 15.6z"
-        className="bg-color"
-      />
-      <path
-        d="M76.5 29.5c-.8-2.8-2.9-5-5.8-5.7-5.1-1.4-25.7-1.4-25.7-1.4s-20.6 0-25.7 1.4c-2.8.7-5 2.9-5.8 5.7-1.4 5-1.4 15.5-1.4 15.5s0 10.6 1.4 15.6c.8 2.7 3 5 5.8 5.7 5.1 1.4 25.7 1.4 25.7 1.4s20.6 0 25.7-1.4c2.9-.8 5-3 5.8-5.7 1.4-5 1.4-15.6 1.4-15.6s0-10.5-1.4-15.5zM38.2 54.7V35.5l17 9.6-17 9.6z"
-        className="icon-color"
-      />
-    </svg>
-  );
-};
-export default React.memo(Youtube);
