+++ apps/mobile/app/ScrollSectionsWrapper.tsx
@@ -1,8 +1,13 @@
 "use client";
+import type { ReactNode } from "react";
 import { useScrollAnchors } from "../src/hooks/useScrollAnchors";
 import { sections } from "../src/assets/data/sections";
 
-const ScrollSectionsWrapper = ({ children }) => {
+type ScrollSectionsWrapperProps = {
+    children: ReactNode;
+};
+
+const ScrollSectionsWrapper = ({ children }: ScrollSectionsWrapperProps) => {
     useScrollAnchors(sections);
     return <>{children}</>;
 };
