+++ apps/mobile/src/components/header/Header.tsx
@@ -5,18 +5,13 @@ import { usePathname } from "next/navigation";
 import Nav from "./Nav";
 import { useScrollContext } from "@utils/context/ScrollContext";
 import { useNavigation } from "@utils/context/NavigationContext";
-import { MenuItem, menuItems } from "@assets/data/menuItems";
+import { menuItems } from "@assets/data/menuItems";
 import { updateMenuClasses } from "@utils/updateMenuUtils";
 import { useSmoothScroll } from "@utils/useSmoothScroll";
 import { useInitialScroll } from "@utils/scrollUtils";
 import { makeClickHandler } from "@utils/handlers";
 
-interface NavProps {
-    menuItems: MenuItem[];
-    onNavigationClick: (path: string) => void;
-}
-
-const Header: React.FC<NavProps> = () => {
+const Header = () => {
     const pathname = usePathname();
     const { currentRoute, updateRoute, closeHamburgerMenu } = useNavigation();
     const { activeSection } = useScrollContext();
