+++ apps/mobile/src/components/header/LogoLink.tsx
@@ -1,8 +1,13 @@
 import { memo } from "react";
+import type { MouseEventHandler } from "react";
 import Link from "next/link";
 import Logo from "../svg_Icon/Logo";
 
-const LogoLink = ({ onClick }) => {
+type LogoLinkProps = {
+    onClick?: MouseEventHandler<HTMLAnchorElement>;
+};
+
+const LogoLink = ({ onClick }: LogoLinkProps) => {
     return (
         <Link href="/#top" aria-label="Vers home page " className="logo-link" onClick={onClick}>
             <Logo />
