+++ apps/mobile/src/components/svg_Icon/Tarifs.tsx
@@ -1,0 +1,14 @@
+import { memo } from "react";
+
+const Tarifs = () => (
+    <div className="nav-icon">
+        <svg viewBox="0 0 18.1 18.1" fill="none" aria-hidden="true">
+            <path
+                d="M11 13.6c.5 0 .8-.4.8-.8s-.4-.8-.8-.8H9.6c-.9 0-1.8-.5-2.3-1.2h2.9a.65.65 0 0 0 .6-.6.65.65 0 0 0-.6-.6H6.8v-.5c0-.2 0-.4.1-.6h3.3a.65.65 0 0 0 .6-.6.65.65 0 0 0-.6-.6H7.3c.5-.7 1.3-1.1 2.3-1.1H11c.5 0 .8-.4.8-.8s-.3-.9-.8-.9H9.6c-1.8 0-3.5 1.1-4.2 2.8h-.9c-.3 0-.5.3-.5.6a.65.65 0 0 0 .6.6h.6v.6.5h-.7 0a.47.47 0 0 0-.5.5.65.65 0 0 0 .6.6h.8c.7 1.7 2.4 2.9 4.2 2.9H11zM9 18a8.96 8.96 0 0 1-9-9 8.96 8.96 0 0 1 9-9 8.96 8.96 0 0 1 9 9 8.96 8.96 0 0 1-9 9zM9 1.6C4.9 1.6 1.6 4.9 1.6 9s3.3 7.4 7.4 7.4 7.4-3.3 7.4-7.4S13.1 1.6 9 1.6z"
+                className="icon-color"
+            />
+        </svg>
+    </div>
+);
+
+export default memo(Tarifs);
